+++ config/config.go
@@ -6,4 +6,5 @@ func Load() {
 	viper.AutomaticEnv()
 	viper.SetDefault("ENVOY_HOST", "http://0.0.0.0:19001")
 	viper.SetDefault("CONSUL_API_HOST", "http://0.0.0.0:8500")
+	viper.SetDefault("NOMAD_ADDR", "https://0.0.0.0:4646")
 }