+++ consul/agent_test.go
@@ -1,33 +1,43 @@
 package consul
 
 import (
-	"bonvoy/test"
-	"net/http"
-	"testing"
+	"bonvoy/config"
 	"github.com/stretchr/testify/require"
+	mock "gopkg.in/h2non/gentleman-mock.v2"
+	"testing"
 )
 
-func AgentLeafCaCertificateMock(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	_, _ = w.Write([]byte(`{
-  "SerialNumber": "01:7e",
-  "CertPEM": "test-cert",
-  "PrivateKeyPEM": "test-pk",
-  "Service": "auth-grpc",
-  "ServiceURI": "spiffe://89749198-098b-d4bc-802b-04d28ed8af0a.consul/ns/default/dc/youngeducated/svc/auth-grpc",
-  "ValidAfter": "2021-06-07T17:48:13Z",
-  "ValidBefore": "2021-06-10T17:48:13Z",
-  "CreateIndex": 271994,
-  "ModifyIndex": 271994
-}`))
-}
-
 func TestClient_GetConnectLeafCaCertificate(t *testing.T) {
-	srv := test.ServerMock("/v1/agent/connect/ca/leaf/auth-grpc", AgentLeafCaCertificateMock)
-	defer srv.Close()
-	client := Client{address: srv.URL}
+	defer mock.Disable()
+	config.Load()
 
-	cert := ConnectLeafCaCertificate{
+	mock.New(GetDefaultAddress()).
+		Get("/v1/agent/connect/ca/leaf/auth-grpc").
+		Reply(200).
+		SetHeader("Content-Type", "application/json").
+		JSON(map[string]interface{}{
+		"SerialNumber": "01:7e",
+		"CertPEM": "test-cert",
+		"PrivateKeyPEM": "test-pk",
+		"Service": "auth-grpc",
+		"ServiceURI": "spiffe://89749198-098b-d4bc-802b-04d28ed8af0a.consul/ns/default/dc/youngeducated/svc/auth-grpc",
+		"ValidAfter": "2021-06-07T17:48:13Z",
+		"ValidBefore": "2021-06-10T17:48:13Z",
+		"CreateIndex": 271994,
+		"ModifyIndex": 271994,
+	})
+
+	client := NewClient()
+	client.client.Use(mock.Plugin)
+	res, err := client.GetConnectLeafCaCertificate("auth-grpc")
+	if err != nil {
+		t.Error(err)
+	}
+	if res.Service != "auth-grpc" {
+		t.Error("Failed to return proper Service name")
+	}
+
+	expectedCert := ConnectLeafCaCertificate{
 		SerialNumber: "01:7e",
 		CertPEM: "test-cert",
 		PrivateKeyPEM: "test-pk",
@@ -38,12 +48,5 @@ func TestClient_GetConnectLeafCaCertificate(t *testing.T) {
 		CreateIndex: 271994,
 		ModifyIndex: 271994,
 	}
-	res, err := client.GetConnectLeafCaCertificate("auth-grpc")
-	if err != nil {
-		t.Error(err)
-	}
-	if res.Service != "auth-grpc" {
-		t.Error("Failed to return proper Service name")
-	}
-	require.Equal(t, cert, res)
+	require.Equal(t, expectedCert, res)
 }